Move MessageInfoHolder to k9mail module
