+++ app/core/src/main/java/com/fsck/k9/helper/MessageHelper.java
@@ -11,12 +11,7 @@ import android.text.style.ForegroundColorSpan;
 import com.fsck.k9.Account;
 import com.fsck.k9.K9;
 import com.fsck.k9.core.R;
-import com.fsck.k9.activity.FolderInfoHolder;
-import com.fsck.k9.activity.MessageInfoHolder;
 import com.fsck.k9.mail.Address;
-import com.fsck.k9.mail.Flag;
-import com.fsck.k9.mail.Message.RecipientType;
-import com.fsck.k9.mailstore.LocalMessage;
 
 public class MessageHelper {
     /**
@@ -45,49 +40,6 @@ public class MessageHelper {
 
     private MessageHelper(final Context context) {
         mContext = context;
-    }
-
-    public void populate(final MessageInfoHolder target,
-                         final LocalMessage message,
-                         final FolderInfoHolder folder,
-                         Account account) {
-        final Contacts contactHelper = K9.showContactName() ? Contacts.getInstance(mContext) : null;
-
-        target.message = message;
-        target.compareArrival = message.getInternalDate();
-        target.compareDate = message.getSentDate();
-        if (target.compareDate == null) {
-            target.compareDate = message.getInternalDate();
-        }
-
-        target.folder = folder;
-
-        target.read = message.isSet(Flag.SEEN);
-        target.answered = message.isSet(Flag.ANSWERED);
-        target.forwarded = message.isSet(Flag.FORWARDED);
-        target.flagged = message.isSet(Flag.FLAGGED);
-
-        Address[] addrs = message.getFrom();
-
-        if (addrs.length > 0 &&  account.isAnIdentity(addrs[0])) {
-            CharSequence to = toFriendly(message.getRecipients(RecipientType.TO), contactHelper);
-            target.compareCounterparty = to.toString();
-            target.sender = new SpannableStringBuilder(mContext.getString(R.string.message_to_label)).append(to);
-        } else {
-            target.sender = toFriendly(addrs, contactHelper);
-            target.compareCounterparty = target.sender.toString();
-        }
-
-        if (addrs.length > 0) {
-            target.senderAddress = addrs[0].getAddress();
-        } else {
-            // a reasonable fallback "whomever we were corresponding with
-            target.senderAddress = target.compareCounterparty;
-        }
-
-        target.uid = message.getUid();
-        target.account = message.getFolder().getAccountUuid();
-        target.uri = message.getUri();
     }
 
     public CharSequence getDisplayName(Account account, Address[] fromAddrs, Address[] toAddrs) {
